avformat/http: avoid int overflow
This was meant to accumulate int64_t timestamp values.

Fixes: b8daba42cdb4b91e793b91ef3055797b0b7611bd

+++ libavformat/http.c
@@ -154,8 +154,8 @@ typedef struct HTTPContext {
     int nb_retries;
     int nb_reconnects;
     int nb_redirects;
-    int sum_latency; /* divide by nb_requests */
-    int max_latency;
+    int64_t sum_latency; /* divide by nb_requests */
+    int64_t max_latency;
 } HTTPContext;
 
 #define OFFSET(x) offsetof(HTTPContext, x)
